fix: Add utility functions to help format date and currency
